+++ app/React.test.tsx
@@ -1,6 +1,6 @@
 import configureStore  from 'redux-mock-store'
 import thunk from 'redux-thunk'
-import {init, getDevicePlatform, getAppVersion} from './React'
+import {init} from './React'
 import {installStore} from './Store'
 import {setDocument, setWindow, setDevice} from './Globals'
 
@@ -93,31 +93,4 @@ describe('React', () => {
     it('logs to google analytics if GA set up');
   })
 
-  describe('getPlatform', () => {
-    it('reports web if no device inititialized', () => {
-      // No initialization at all
-      expect(getDevicePlatform()).toEqual('web');
-    });
-
-    it('defaults to web on unexpected device', () => {
-      setDevice({platform: 'zune'});
-      expect(getDevicePlatform()).toEqual('web');
-    });
-
-    it('reports ios if ios device initialized', () => {
-      setDevice({platform: 'ios'});
-      expect(getDevicePlatform()).toEqual('ios');
-    });
-
-    it('reports android if android device initialized', () => {
-      setDevice({platform: 'android'});
-      expect(getDevicePlatform()).toEqual('android');
-    });
-  });
-
-  describe('getVersion', () => {
-    it('gets a version string', () => {
-      expect(getAppVersion()).toMatch(/[0-9]+\.[0-9]+\.[0-9]+/);
-    });
-  });
 });
