+++ app/React.tsx
@@ -10,14 +10,9 @@ import {authSettings} from './Constants'
 import {toPrevious} from './actions/Card'
 import {silentLogin} from './actions/User'
 import {getStore} from './Store'
-import {getWindow, getGapi, getGA, getDevice, getDocument, setGA} from './Globals'
+import {getWindow, getGapi, getGA, getDevicePlatform, getDocument, setGA} from './Globals'
 
-const PACKAGE = require('../package.json');
 const injectTapEventPlugin = require('react-tap-event-plugin');
-
-export function getAppVersion(): string {
-  return PACKAGE.version;
-}
 
 function setupTapEvents() {
   try {
@@ -27,22 +22,7 @@ function setupTapEvents() {
   }
 }
 
-export function getDevicePlatform(): 'android' | 'ios' | 'web' {
-  const device = getDevice();
 
-  if (device === undefined) {
-    return 'web';
-  }
-
-  var p = (device.platform || '').toLowerCase();
-  if (/android/i.test(p)) {
-    return 'android';
-  } else if (/iphone|ipad|ipod|ios/i.test(p)) {
-    return 'ios';
-  } else {
-    return 'web';
-  }
-}
 
 export function logEvent(name: string, args: any): void {
   getWindow().FirebasePlugin.logEvent(name, args);
@@ -55,14 +35,9 @@ export function logEvent(name: string, args: any): void {
 
 function setupDevice() {
   const window = getWindow();
-  const device = getDevice();
 
-  // This is used to mutate device properties for tests
-  window.device = device;
-
-  var platform = getDevicePlatform();
-
-  window.platform = platform; // TODO: remove this and have everyone use getPlatform().
+  // Apply class-specific styling
+  const platform = getDevicePlatform();
   document.body.className += ' ' + platform;
 
   if (platform === 'android') {
@@ -198,7 +173,6 @@ export function init() {
   // TODO: remove these and have everyone use getPlatform()/getVersion().
   const window = getWindow();
   window.platform = 'web';
-  window.APP_VERSION = getAppVersion();
 
   getDocument().addEventListener('deviceready', () => {
     setupDevice();
